faq-capture-bot/requirements.txt
rasa==2.3.4
rasa-sdk==2.3.1
spacy==2.3.4
requests==2.25.0
slack-sdk==3.4.2
python-dotenv==0.15.0